refactor: split crud.py into domain modules
Extracts all logic from the 1667-line crud.py monolith into:
- db/conversations.py  — agent/conversation/proposals/threads
- db/organizations.py  — org/user/member management
- db/themes.py         — project/theme/version + decisions
- db/knowledge.py      — factor/claim CRUD + ID resolution

crud.py becomes a re-export facade; all existing imports remain
compatible. 47 auth coverage tests pass without regression.

Closes #284
